Initial C_wrapping logic with builder script.
